Earthquakes!
Studying earthquake inter-arrival times, magnitudes and occurrences led us to generally conclude a number of possible
”simplified” distributions to predict them. For one, we believe that the inter-arrival times appear to fit the beta or gamma
distribution the best,instead of the ideal standard which would have been the exponential distribution - chalked up to the lack of
independence between earthquakes and aftershocks. Earthquake occurrences were generally indicative of Poisson distributions
when considering a relatively low number of days (100 days) and got superseded by the negative binomial distribution due to
the inflated presence of days with 0 events for larger periods (500 days).